+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.7.1 / 2018-04-23
+==================
+
+  * Request mocks
+  * Browser.style
+  * Components improved, race condition warning fixed
+
 0.7.0 / 2018-04-20
 ==================